split timeline code into a separate file
